requirements.txt
fastapi
uvicorn
openai
tiktoken
regex
python-dotenv
diskcache
pypdf
python-multipart
jinja2
